requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
django-allauth==65.10.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-auth-kit==0.3.10
drf-spectacular==0.29.0
hashids==1.3.1
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
oauthlib==3.3.1
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
pyotp==2.9.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
sqlparse==0.5.3
structlog==25.5.0
uritemplate==4.2.0
urllib3==2.5.0
